% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMS-package.R
\docType{package}
\name{missMS-package}
\alias{missMS-package}
\alias{missMS}
\title{Missing data models for label-free proteomics}
\description{
The \pkg{missMS} package enables users to fit missing data models
models for label-free mass spectrometry experiments.  The package is
designed to make these models accesible to people without a background
in Bayesian statistics or missing data modeling.  This is done by
requiring users to follow a very specific format for data entry.  Once
the data is formatted properly everything else is handled internally.
The required fields and headers are explained in the documentation for the
sample data NEED SAMPLE DATA.
}
\section{Model Details}{

This is where I will explain the model and the options for use:
\describe{
Here is a detailed description.
}
}

