% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resn}
\alias{resn}
\title{Generate a random number from an extended skew normal distribution}
\usage{
resn(n_, xi, omega, alpha, tau)
}
\arguments{
\item{n_}{The number of random variates to create}

\item{xi}{esn parameter}

\item{omega}{esn parameter}

\item{alpha}{esn parameter}

\item{tau}{esn parameter}
}
\description{
Generate a random number from an extended skew normal distribution
}
