% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuns.R
\name{testfun}
\alias{testfun}
\title{Generate a random number for Gaussian mean parameter}
\usage{
testfun(n_, xi)
}
\arguments{
\item{n_}{The number of random variates to create}

\item{xi}{esn parameter}

\item{omega}{esn parameter}

\item{alpha}{esn parameter}

\item{tau}{esn parameter}
}
\description{
Generate a random number for Gaussian mean parameter
}
