% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smp}
\alias{smp}
\title{Fit the Selection Model for Proteomics}
\usage{
smp(dat, nCores = 1, ndraws = 2000)
}
\arguments{
\item{dat}{Correctly formated and normalized data frame}

\item{nCores}{Number of cores to use}

\item{ndraws}{Number of draws of the Gibbs Sampler}
}
\description{
Fit the Selection Model for Proteomics
}
